+++ src/components/my-map/index.ts
@@ -1,4 +1,4 @@
-import { css, html, LitElement } from "lit";
+import { html, LitElement, unsafeCSS } from "lit";
 import { customElement, property } from "lit/decorators.js";
 import { Control, defaults as defaultControls } from "ol/control";
 import { GeoJSON } from "ol/format";
@@ -7,53 +7,42 @@ import { Vector as VectorLayer } from "ol/layer";
 import Map from "ol/Map";
 import { fromLonLat, transformExtent } from "ol/proj";
 import { Vector as VectorSource } from "ol/source";
-import { Stroke, Style } from "ol/style";
+import { Fill, Stroke, Style } from "ol/style";
 import View from "ol/View";
 import { last } from "rambda";
 
-import { draw, drawingLayer, drawingSource, modify, snap } from "./draw";
-import { scaleControl } from "./scale-line"
 import {
+  configureDraw,
+  configureModify,
+  drawingLayer,
+  drawingSource,
+  DrawPointerEnum,
+  snap,
+} from "./drawing";
+import {
+  getFeaturesAtPoint,
   makeFeatureLayer,
   outlineSource,
-  getFeaturesAtPoint,
 } from "./os-features";
 import { makeOsVectorTileBaseMap, makeRasterBaseMap } from "./os-layers";
-import { AreaUnitEnum, fitToData, formatArea } from "./utils";
+import { scaleControl } from "./scale-line";
+import {
+  getSnapPointsFromVectorTiles,
+  pointsLayer,
+  pointsSource,
+} from "./snapping";
+import { AreaUnitEnum, fitToData, formatArea, hexToRgba } from "./utils";
+import styles from "./styles.scss";
 
 @customElement("my-map")
 export class MyMap extends LitElement {
-  // default map size, can be overridden with CSS
-  static styles = css`
-    :host {
-      display: block;
-      width: 800px;
-      height: 800px;
-      position: relative;
-    }
-    #map {
-      height: 100%;
-      opacity: 0;
-      transition: opacity 0.25s;
-      overflow: hidden;
-    }
-    #map:focus {
-      outline: #424242 solid 0.15em;
-    }
-    .reset-control {
-      top: 70px;
-      left: 0.5em;
-    }
-    #area {
-      position: absolute;
-      bottom: 0;
-      left: 0;
-      z-index: 100;
-      background: white;
-    }
-  `;
+  // ref https://github.com/e111077/vite-lit-element-ts-sass/issues/3
+  static styles = unsafeCSS(styles);
 
   // configurable component properties
+  @property({ type: String })
+  id = "map";
+
   @property({ type: Number })
   latitude = 51.507351;
 
@@ -72,6 +61,18 @@ export class MyMap extends LitElement {
   @property({ type: Boolean })
   drawMode = false;
 
+  @property({ type: Object })
+  drawGeojsonData = {
+    type: "Feature",
+    geometry: {},
+  };
+
+  @property({ type: Number })
+  drawGeojsonDataBuffer = 100;
+
+  @property({ type: String })
+  drawPointer: DrawPointerEnum = "crosshair";
+
   @property({ type: Boolean })
   showFeaturesAtPoint = false;
 
@@ -80,6 +81,9 @@ export class MyMap extends LitElement {
 
   @property({ type: String })
   featureColor = "#0000ff";
+
+  @property({ type: Boolean })
+  featureFill = false;
 
   @property({ type: Number })
   featureBuffer = 40;
@@ -93,6 +97,9 @@ export class MyMap extends LitElement {
   @property({ type: String })
   geojsonColor = "#ff0000";
 
+  @property({ type: Boolean })
+  geojsonFill = false;
+
   @property({ type: Number })
   geojsonBuffer = 12;
 
@@ -105,6 +112,9 @@ export class MyMap extends LitElement {
   @property({ type: String })
   osFeaturesApiKey = import.meta.env.VITE_APP_OS_FEATURES_API_KEY || "";
 
+  @property({ type: String })
+  osCopyright = `© Crown copyright and database rights ${new Date().getFullYear()} OS (0)100024857`;
+
   @property({ type: Boolean })
   hideResetControl = false;
 
@@ -112,26 +122,35 @@ export class MyMap extends LitElement {
   staticMode = false;
 
   @property({ type: String })
-  areaUnit: AreaUnitEnum = "m2"
+  areaUnit: AreaUnitEnum = "m2";
 
-  @property({ type: String })
-  ariaLabel = "Interactive map";
-  
   @property({ type: Boolean })
   showScale = false;
 
   @property({ type: Boolean })
   useScaleBarStyle = false;
 
+  // set class property (map doesn't require any reactivity using @state)
+  map?: Map;
+
+  // called when element is created
+  constructor() {
+    super();
+  }
+
   // runs after the initial render
   firstUpdated() {
-    const target = this.shadowRoot?.querySelector("#map") as HTMLElement;
+    const target = this.renderRoot.querySelector(`#${this.id}`) as HTMLElement;
 
     const useVectorTiles =
       !this.disableVectorTiles && Boolean(this.osVectorTilesApiKey);
 
-    const rasterBaseMap = makeRasterBaseMap(this.osVectorTilesApiKey);
+    const rasterBaseMap = makeRasterBaseMap(
+      this.osCopyright,
+      this.osVectorTilesApiKey
+    );
     const osVectorTileBaseMap = makeOsVectorTileBaseMap(
+      this.osCopyright,
       this.osVectorTilesApiKey
     );
 
@@ -163,6 +182,12 @@ export class MyMap extends LitElement {
       }),
     });
 
+    this.map = map;
+
+    // make configurable interactions available
+    const draw = configureDraw(this.drawPointer);
+    const modify = configureModify(this.drawPointer);
+
     // add a custom 'reset' control below zoom
     const button = document.createElement("button");
     button.innerHTML = "↻";
@@ -180,11 +205,20 @@ export class MyMap extends LitElement {
 
       if (this.drawMode) {
         drawingSource.clear();
+
+        this.dispatch("geojsonChange", {});
+        this.dispatch(
+          "areaChange",
+          `0 ${this.areaUnit === "m2" ? "m²" : this.areaUnit}`
+        );
+
         map.addInteraction(draw);
         map.addInteraction(snap);
       }
     };
 
+    // this is an internal event listener, so doesn't need to be removed later
+    // ref https://lit.dev/docs/components/lifecycle/#disconnectedcallback
     button.addEventListener("click", handleReset, false);
 
     const element = document.createElement("div");
@@ -197,6 +231,13 @@ export class MyMap extends LitElement {
       map.addControl(ResetControl);
     }
 
+    // Apply aria-labels to OL Controls for accessibility
+    const olControls: NodeListOf<HTMLButtonElement> | undefined =
+      this.renderRoot?.querySelectorAll(".ol-control button");
+    olControls?.forEach((node) =>
+      node.setAttribute("aria-label", node.getAttribute("title") || "")
+    );
+
     // define cursors for dragging/panning and moving
     map.on("pointerdrag", () => {
       map.getViewport().style.cursor = "grabbing";
@@ -206,7 +247,7 @@ export class MyMap extends LitElement {
       map.getViewport().style.cursor = "grab";
     });
 
-    // add a vector layer to display static geojson if features are provided
+    // display static geojson if features are provided
     const geojsonSource = new VectorSource();
 
     if (this.geojsonData.type === "FeatureCollection") {
@@ -228,6 +269,11 @@ export class MyMap extends LitElement {
           color: this.geojsonColor,
           width: 3,
         }),
+        fill: new Fill({
+          color: this.geojsonFill
+            ? hexToRgba(this.geojsonColor, 0.2)
+            : hexToRgba(this.geojsonColor, 0),
+        }),
       }),
     });
 
@@ -237,26 +283,46 @@ export class MyMap extends LitElement {
       // fit map to extent of geojson features, overriding default zoom & center
       fitToData(map, geojsonSource, this.geojsonBuffer);
 
-      // log total area of first feature (assumes geojson is a single polygon for now)
+      // log total area of static geojson data (assumes single polygon for now)
       const data = geojsonSource.getFeatures()[0].getGeometry();
-      console.log("geojsonData total area:", formatArea(data, this.areaUnit));
+      if (data) {
+        this.dispatch("geojsonDataArea", formatArea(data, this.areaUnit));
+      }
     }
 
+    // draw interactions
     if (this.drawMode) {
-      // ensure we start from an empty array of features
+      // make sure drawingSource is cleared upfront, even if drawGeojsonData is provided
       drawingSource.clear();
+
+      // load an initial polygon into the drawing source if provided, or start from an empty drawing source
+      const loadInitialDrawing =
+        Object.keys(this.drawGeojsonData.geometry).length > 0;
+
+      if (loadInitialDrawing) {
+        let feature = new GeoJSON().readFeature(this.drawGeojsonData, {
+          featureProjection: "EPSG:3857",
+        });
+        drawingSource.addFeature(feature);
+        fitToData(map, drawingSource, this.drawGeojsonDataBuffer);
+      }
+
       map.addLayer(drawingLayer);
 
-      map.addInteraction(draw);
-      map.addInteraction(snap);
+      if (!loadInitialDrawing) {
+        map.addInteraction(draw);
+      }
       map.addInteraction(modify);
+
+      // XXX: snap must be added after draw and modify
+      map.addInteraction(snap);
 
       // 'change' listens for 'drawend' and modifications
       drawingSource.on("change", () => {
         const sketches = drawingSource.getFeatures();
 
         if (sketches.length > 0) {
-          const lastSketchGeom = last(sketches).getGeometry();
+          const lastSketchGeom = last(sketches)?.getGeometry();
 
           this.dispatch(
             "geojsonChange",
@@ -265,28 +331,74 @@ export class MyMap extends LitElement {
             })
           );
 
-          this.dispatch("areaChange", formatArea(lastSketchGeom, this.areaUnit));
+          if (lastSketchGeom) {
+            this.dispatch(
+              "areaChange",
+              formatArea(lastSketchGeom, this.areaUnit)
+            );
+          }
 
           // limit to drawing a single polygon, only allow modifications to existing shape
           map.removeInteraction(draw);
-          map.removeInteraction(snap);
         }
       });
     }
 
+    // show snapping points when in drawMode, with vector tile basemap enabled, and at qualifying zoom
+    if (
+      this.drawMode &&
+      Boolean(this.osVectorTilesApiKey) &&
+      !this.disableVectorTiles
+    ) {
+      // define zoom threshold for showing snaps (not @property yet because computationally expensive!)
+      const snapsZoom: number = 20;
+
+      // display draw vertices on top of snap points
+      map.addLayer(pointsLayer);
+      drawingLayer.setZIndex(1001);
+
+      // extract snap-able points from the basemap, and display them as points on the map if initial render within zoom
+      if (this.zoom < snapsZoom) {
+        pointsSource.clear();
+        const extent = map.getView().calculateExtent(map.getSize());
+        getSnapPointsFromVectorTiles(osVectorTileBaseMap, extent);
+      }
+
+      // continue to fetch & update snaps as map moves
+      map.on("moveend", () => {
+        const currentZoom: number | undefined = map.getView().getZoom();
+        if (currentZoom && currentZoom < snapsZoom) {
+          pointsSource.clear();
+          return;
+        }
+
+        // timeout minimizes snap updates mid-pan/drag
+        setTimeout(() => {
+          pointsSource.clear();
+          const extent = map.getView().calculateExtent(map.getSize());
+          getSnapPointsFromVectorTiles(osVectorTileBaseMap, extent);
+        }, 200);
+      });
+    }
+
+    // OS Features API & click-to-select interactions
     if (this.showFeaturesAtPoint && Boolean(this.osFeaturesApiKey)) {
       getFeaturesAtPoint(
         fromLonLat([this.longitude, this.latitude]),
-        this.osFeaturesApiKey
+        this.osFeaturesApiKey,
+        false
       );
 
       if (this.clickFeatures) {
         map.on("singleclick", (e) => {
-          getFeaturesAtPoint(e.coordinate, this.osFeaturesApiKey);
+          getFeaturesAtPoint(e.coordinate, this.osFeaturesApiKey, true);
         });
       }
 
-      const outlineLayer = makeFeatureLayer(this.featureColor);
+      const outlineLayer = makeFeatureLayer(
+        this.featureColor,
+        this.featureFill
+      );
       map.addLayer(outlineLayer);
 
       // ensure getFeaturesAtPoint has fetched successfully
@@ -298,9 +410,22 @@ export class MyMap extends LitElement {
           // fit map to extent of features
           fitToData(map, outlineSource, this.featureBuffer);
 
-          // log total area of feature or merged features
+          // write the geojson representation of the feature or merged features
+          this.dispatch(
+            "featuresGeojsonChange",
+            new GeoJSON().writeFeaturesObject(outlineSource.getFeatures(), {
+              featureProjection: "EPSG:3857",
+            })
+          );
+
+          // calculate the total area of the feature or merged features
           const data = outlineSource.getFeatures()[0].getGeometry();
-          console.log("feature(s) total area:", formatArea(data, this.areaUnit));
+          if (data) {
+            this.dispatch(
+              "featuresAreaChange",
+              formatArea(data, this.areaUnit)
+            );
+          }
         }
       });
     }
@@ -317,10 +442,13 @@ export class MyMap extends LitElement {
   render() {
     return html`<script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
       <link rel="stylesheet" href="https://cdn.skypack.dev/ol@^6.6.1/ol.css" />
-      <div 
-        id="map" 
-        tabindex="0"
-        aria-label=${this.ariaLabel} />`;
+      <div id="${this.id}" class="map" tabindex="0" />`;
+  }
+
+  // unmount the map
+  disconnectedCallback() {
+    super.disconnectedCallback();
+    this.map?.dispose();
   }
 
   /**